feat: all types add in types file and fetch the data in dashboard
