+++ src/app/constants/constants.ts
@@ -1,1 +1,3 @@
 export const ATTENDANCE_CUTOFF_TIME = '10:00';
+
+export type AttendanceStatus = "ON_TIME" | "LATE" | "ABSENT";
